Move form components to subfolder
