+++ src/pages/Today.js
@@ -6,7 +6,7 @@ import { Helmet } from 'react-helmet-async';
 // components
 import Layout from 'components/Layout';
 import TaskList from 'components/TaskList';
-import TaskCreateForm from 'components/Forms/TaskCreateForm';
+import TaskCreateForm from 'components/forms/TaskCreateForm';
 
 // store
 import { useStore } from 'store/useStore';
